Geo coder

+++ cllc-interfaces/Dynamics-Autorest/Models/MicrosoftDynamicsCRMDoubleAttributeMetadata.cs
@@ -29,7 +29,7 @@ namespace Gov.Lclb.Cllb.Interfaces.Models
         /// </summary>
         /// <param name="imeMode">Possible values include: 'Auto', 'Inactive',
         /// 'Active', 'Disabled'</param>
-        public MicrosoftDynamicsCRMDoubleAttributeMetadata(string imeMode = default(string), double? maxValue = default(double?), double? minValue = default(double?), int? precision = default(int?))
+        public MicrosoftDynamicsCRMDoubleAttributeMetadata(string imeMode = default(string), decimal? maxValue = default(decimal?), decimal? minValue = default(decimal?), int? precision = default(int?))
         {
             ImeMode = imeMode;
             MaxValue = maxValue;
@@ -53,12 +53,12 @@ namespace Gov.Lclb.Cllb.Interfaces.Models
         /// <summary>
         /// </summary>
         [JsonProperty(PropertyName = "MaxValue")]
-        public double? MaxValue { get; set; }
+        public decimal? MaxValue { get; set; }
 
         /// <summary>
         /// </summary>
         [JsonProperty(PropertyName = "MinValue")]
-        public double? MinValue { get; set; }
+        public decimal? MinValue { get; set; }
 
         /// <summary>
         /// </summary>
